edited styling
