+++ src/components/LoginPage.module.css
@@ -1,0 +1,5 @@
+.box {
+  display: flex;
+  justify-content: flex-end;
+  gap: 5rem;
+}
