+++ src/components/PageContent.module.css
@@ -3,6 +3,7 @@
   background-color: #ecf9f2;
   padding: 4.8rem 0 9.6rem 0;
   margin: 0 auto;
+  margin-top: 9.6rem;
   padding: 0 3.2rem;
   display: grid;
   grid-template-rows: 1fr 2fr 1fr;
